refactor: 용도에 따른 package 분리
