+++ src/main/java/carracing/cars/StepForwardStrategy.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.cars;
 
 public interface StepForwardStrategy {
     boolean stepForward();
