+++ src/main/java/carracing/utils/RandomFactory.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.utils;
 
 import java.util.Random;
