+++ src/main/java/carracing/view/InputView.java
@@ -1,4 +1,6 @@
-package carracing;
+package carracing.view;
+
+import carracing.cars.CarName;
 
 import java.util.Arrays;
 import java.util.List;
