+++ src/test/java/carracing/CarNameTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.cars.CarName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
