+++ src/test/java/carracing/CarTest.java
@@ -1,5 +1,7 @@
 package carracing;
 
+import carracing.cars.Car;
+import carracing.cars.CarName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
