+++ src/test/java/carracing/RandomFactoryTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.utils.RandomFactory;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
